# Use std::popcount instead of __builtin_popcount